Update some workers as CPU bound
The following workers are found to be CPU bounded:

* ExternalServiceReactiveCachingWorker
* MergeRequests::MergeabilityCheckBatchWorker
* ContainerRegistry::DeleteContainerRepositoryWorker
* ContainerExpirationPolicies::CleanupContainerRepositoryWorker
* Security::ScanExecutionPolicies::CreatePipelineWorker

So this updates their `worker_resource_boundary` attribute to `:cpu` so
that they could be processed by the dedicated CPU bound shard, ie
`low-urgency-cpu-bound` shard in GitLab.com. It also helps to alleviate
the sidekiq thread contention in the catchall shard.