Make Coverband work properly with Puma processes
Previously code paths hitting Puma was not being reported at all by
Coverband. This was not working for two main reasons:

1. Normally Coverband expects that the gem is loaded at boot time,
which registers a Railtie that inserts a middleware into the Rails
stack. This middleware spawns a background thread that reports code
coverage.

However, in GitLab, the Coverband gem is dynamically required in
`config/initializers`. Because the gem was loaded late, this
middleware was never being inserted. GitLab loads Coverband this way
because the Redis shared state configuration in set up in
`config/initializers/7_redis.rb`.

Since the middleware stack is already frozen here, we can work around
this by manually starting the background thread.

However, we also have to be aware that Puma launches multiple worker
processes. Each worker process needs to have its own background thread
to report code hits from the process. We use the Puma
`on_worker_start` lifecycle hook to do this.

2. The Coverband root directory must be set properly for application
code to be recognized. Previously the root directory was set to
`Rails.root('tmp')` because Puma starts in a temporary directory to
model the behavior of Omnibus GitLab. This commit forces the root
directory to be `Rails.root`.

Relates to
https://gitlab.com/gitlab-org/quality/quality-engineering/team-tasks/-/issues/3532