Refactor "Change type" modal supportedConversionTypes format
In the modal component, the array of conversion types is mapped to a
`gl-form-select`-compatible `{ text, value }` format. However,
this is passed around and adds confusion because we expect the array
to be in the untouched `{ id, name }` format. We are therefore
updating the code to always be in the `{ id, name }` format, and
convert to a `{ text, value }` format for `gl-form-select` only.