Enforce dynamic feature flag rubocop rule for Ai Gateway
We recently
[introduced a RuboCop rule](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/186357)
to prevent people from using dynamic keys when checking if a feature
flag is enabled.

We also support pushing feature flags to the AI Gateway (see
https://docs.gitlab.com/development/ai_features/#push-feature-flags-to-ai-gateway)
which uses a different syntax. For the same reasons we want to block the
usage of dynamic keys when using this approach.