Add possibility to hide Duo Base from Duo frontend
Currently Duo Base is under active development, we want to have the
possibility to disable the support for the new add-on in the frontend.
With that we can ensure existing E2E are running through and backfilling
add-on purchases don't give yet the possibility to manage the new
add-on.

EE: true
Changelog: added