Adds step to reconfigure and start GitLab after the restore
For https://gitlab.com/gitlab-org/gitlab/-/issues/440457

In https://gitlab.com/gitlab-com/gl-infra/data-access/durability/team/-/issues/120
while testing PG16 restores we saw that after restore pgbouncer does not have the proper ownership
of pg_shadow_lookup to conduct its authentication. To resolve this we can run a `gitlab-ctl reconfigure`
on the Posgres nodes after the restore is completed.

Instead of a restart, this adds a start because the services were
stopped earlier and restarting is not necessary.