Create Japanese documentation directory with 1 translated file
Adds a translated version of the CI/CD YAML reference documentation
(https://docs.gitlab.com/ci/yaml/) as the first file in the new
/doc-locale/ja-jp/ directory.

This new directory will house all Japanese translated documentation.
While this change exceeds size limitations due to the comprehensive
nature of the translation, adding the complete file in a single commit
is necessary to establish the proper directory structure and maintain
document integrity.

Related to the Japanese localization initiative,
https://gitlab.com/groups/gitlab-com/localization/-/epics/14