Also skip locking tables with disallow_database_ddl_feature_flags
We have the feature flag `disallow_database_ddl_feature_flags` that is
commonly used during database maintenance and is especially useful
during upgrades where we are relying on logical replication. The point
of this is to avoid surprising DDL schema changes happening while it
enabled.

We have `::Database::MonitorLockedTablesWorker` which is a
worker that is intended to add triggers to the database for new tables
to lock them being written on the wrong databases (e.g. locking CI
tables on Main databases). This worker runs on a cron schedule every 3
days.

Considering that it makes schema changes (adding triggers) and "every 3
days" is likely to lead to surprises if we were doing an upgrade then it
makes sense to also skip the schema changes from happening in this
worker if `disallow_database_ddl_feature_flags` is enabled.

I also considered putting this feature flag check inside of the
`LockTablesWorker` class so that it couldn't make schema changes even if
run directly (e.g. via rake task) but that would conflict with intended
uses during decompositions like
https://ops.gitlab.net/gitlab-com/gl-infra/production/-/issues/12 where
we actually want to lock the tables during the maintenance window.