Fix MR diffs endpoint to respond with pagination headers
We previously fixed an issue with MR diffs REST API endpoint that
makes it error out when `per_page` param is higher than the maximum
`per_page` we support.

The fix was to remove the `paginate` call in code. As a side-effect
though, it removes the pagination headers from the response.

To fix that without re-introducing the bug we previously fixed, we
now modify the `paginate` helper to accept a `skip_pagination_check`
param (defaults to `false`). We set it to `true` in this endpoint
because we are already paginating in `#paginated_diffs`.

Changelog: fixed