Refactor & Optimise Ci::Secrets::Integration class
Prior to this change, in the
Ci::Secrets::Integration class, the
Ci::Variables::Collection instance was converted to a hash and then
to check for a specific value, we were searching for the value
in the hash. This is completely unnecessary and inefficient as the
Ci::Variables::Collection class already has a method to fetch a value
by key. By converting to hash, we are wasting memory and CPU cycles.

This change fixes the issue by reusing the
existing method thereby avoiding converting it into a hash and then
searching the hash.

Issue
https://gitlab.com/gitlab-org/gitlab/-/issues/535399