Use CSS-based loaders for work item loading state
the `GlSkeletonLoader` component is not responsive; it will scale in
height and width when stretched to fill its parent. That means that
loaders end up looking very zoomed in when the viewport is wide and the
user uses the "fluid" layout option.

This MR replaces the `GlSkeletonLoader` components used in work item
loading states with the recommended CSS-based loader pattern, which can
be styled with tailwind utils to beter match the responsive layout of a
work item.

Changelog: fixed