Do not check if a user can read_build when returning details_path
The frontend already uses the value of has_details? to correctly display
a build's status information so there is no need for the backend to
return nil for details_path when the user has no read_build ability.

Without reverting this change we prevent the workaround for the bug
described in https://gitlab.com/gitlab-org/gitlab/-/issues/427969.