Add a todo when a merge request review is submitted
Updates the todo service to create a new todo for the author and all assignees
of the merge request whenever a new review is submitted.
This will also be done when GitLab Duo finishes its review, it will create a new
todo for the author and assignees of the merge request.

https://gitlab.com/gitlab-org/gitlab/-/issues/536724