Place SkipJobs middleware before LoadBalancing
During incident, we found that skipping jobs could take significant
amount of time, averaging 1s up to 26s. Majority of these were spent in
quering WALs in the sidekiq server middleware
(`Gitlab::Database::LoadBalancing::SidekiqServerMiddleware`) before the
SkipJobs are executed.

The WAL queries are obsolete at this point if the job ends up being
deferred because the WAL queries still need to be executed again
when the job is resumed for execution.

This reorders the SkipJobs middleware to be executed before
LoadBalancing.