refactor: Avoid escaping characters in error message in package overview
When an errorneous package is pushed to the pckage registry,
the error message is displayed in the package overview.
The original implementation escaped characters like
quotes and backslashes in the error message,
which made it difficult to read.
This was also noticed and mentioned here: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/184059#note_2458186537

This commit modifies the code to display the error message
without escaping characters.

Changelog: other