Updated merge requests query to support review states OR no reviewer
Updated the merge request GraphQL query to allow for filtering merge requests
that have no reviewer OR reviewers with the passed in state.
This will allow us to display lists on the merge request homepage that will show
the user all their assigned merge requests that are currently active ie.
that have no reviewer assigned (draft work) OR where a reviewer has reviewed.