Remove development-only behaviour change
Removes a change in behaviour for development environment, where
integrations that are disabled for the project will not appear as
disabled.

This causes confusion for developers who observe difference between
their environment and production, or test.

If developers want to see all integrations, including disabled ones,
they can modify their codebase locally to do so.