Add author_email column to ssh_signatures table
UI-signed commits are verified using a user_id.
However, this creates a problem because UI commits
are signed by GitLab itself, not a real user.
When verification relies on a user_id,
it can lead to incorrect results.

To address this, we are adding an author_email column
to the ssh_signatures table. This will allow us
to verify UI-signed commits based on the author’s email,
rather than relying on user_id.
The logic changes will be implemented in a separate MR
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/189034