Ensure production code uses axios util
We have wrapped the axios library with some custom interceptors that
help E2E tests, show a modal in case a captcha needs to be shown, etc.

However, a few bits of the code base are not using that wrapper and
import axios directly. In order to ensure a consistent behaviour in our
apps, let all of our client apps consume the shared util.

Also adding an eslint rule to prevent future direct imports.