Prevent unauthorized users from adding, selecting or archiving designs
When a user does not have permission to add designs, prevent them from
seeing the "Add" button in the work item design widget.

When the user does not have permission to update designs, prevent them
from seeing the the "Archive" button, the "Archive all" button, and the
checkbox to select designs.

When a user does not have permission to update designs, also prevent
them from seeing the archive button on a design detail modal.

If a user attempts these actions today while they do not have permission
then the action will fail, but we should not show these affordances when
the user does not have permission.

Changelog: fixed