Ensure primary button on board form remains disabled before modal closes
Currently, the primary button on a board form modal (the button used to
either create, save changes to, or delete a board) is only disabled
while loading, and loading ceases before the modal closes on a success.
This means that the button will briefly stop loading and look like it is
in a clickable state just before the modal closes.

This change adds a new boolean property `isDisabled` that can keep the
button disabled after loading stops before the modal closes.

Changelog: fixed