Update fast_gettext gem to v4.1.0
We're seeing a warning now:

```
/opt/gitlab/embedded/lib/ruby/gems/3.3.0/gems/fast_gettext-2.3.0/lib/fast_gettext/vendor/mofile.rb:24:
warning: prime was loaded from the standard library, but is not part
of the default gems starting from Ruby 3.1.0.  You can add prime to
your Gemfile or gemspec to silence this warning.  Also please contact
the author of fast_gettext-2.3.0 to request adding prime into its
gemspec.
```

This was fixed in fast_gettext v2.4.0:
https://github.com/grosser/fast_gettext/pull/140

Changelog: fixed