Remove duo_additional_context feature flag
Remove the feature flag `duo_additional_context` that was used to enable
the ability to include additional context into Duo Chat questions.

Removal of this feature flag is part of the cleanup process of the
feature introduction. This feature flag did require at least one of the
other additional feature flags that allowed for specific context to be
enabled for it to affect any users.

EE: true
Changelog: added