Hide list of projects assigned to runner when data is unavailable
Some users (e.g. user with `read_admin_cicd` custom admin role) may not
have read access to all projects assigned to a project runner. In this
scenario runner.projects.nodes will be empty but runner.projectCount > 0
.

In this case we hide the projects list to prevent confusion (e.g.
displayed count is > 0 but there are no projects in the list). Instead,
the user can go to the jobs tab to view builds of projects assigned to
the runner.