Fix CI depended triggers not locks resources
**Problem:**

When a trigger job's status is retried via downstream pipeline,
it does not hold its `resource_group` again.

**Workflow:**

- `Ci::RetryJobService` calls `Ci::ResetSkippedJobsService`
- `Ci::ResetSkippedJobsService` calls
`Ci::Pipeline#reset_source_bridge!`
- `reset_source_bridge!` calls `Ci::EnqueueJobService` with the
`pending!` transition parameter.

This directly makes the trigger job "pending".

**Fix:**

First change is that we remove this custom transition and use the
default `enqueue!`. The second change is that we change the status of
the trigger job to "created".

Changelog: fixed