Remove unnecessary dependency exclude list from workhorse
It seems like in the past we used the `exclude` directive in `go.mod` to
yank dependencies from the dependency tree which were vulnerable and not
required directly. However, none of the listed dependencies still seem
part of our transitive dependency tree.

Renovate shined a light on this when it tried to update the
dependencies, which doesn't make sense (and is an upstream issue):
- https://github.com/renovatebot/renovate/discussions/31905