Implement circuitbreaker/roundTripper using gobreaker
The first of a series of MRs to add a circuitbreaker to workhorse.
This circuitbreaker aims to reduce load on rails when the user has been
rate limited. If the user receives a set number of consecutive 429
responses from rails, workhorse will immediately respond with a 429
instead of forwarding the request to rails.

Uses workhorse's existing Redis connection to persist the circuitbreaker
state.