Reuse existing ApprovalMergeRequestRulePolicy
For now, we only want to replicate the existing policy behavior. Rather
than keeping two sources of truth for this, we can declare a policy in
the new class that points to the existing policy.

Once we have a need that diverges from the existing policy, we can
create a new policy class.