Truncate p_ci_build_trace_metadata partition 102 on .com
The table 'p_ci_build_trace_metadata' is used for emphemeral
data to track archival attempts on job traces. After it
successfully archives a trace, we no longer need the metadata
record. We recently implemented a process to destroy the record
after archival, and now we need to delete existing records of
successful archives.

We determined that the best way to do this is to truncate
the old partition. We've recently created and enabled traffic on
partition 103 of p_ci_build_trace_metadata.
This means the traffic on 102 is now very low
or negligible and so we can truncate the existing
data on it safely. Truncating it will allow us
to reclaim ~200 Gb of disk space. We also determined
that unsuccessfully archived
traces will simply be retried by ArchiveTracesCronWorker
as needed.

This truncation applies to Gitlab.com only.

Changelog: other