Fix read AI Setting error when not logged in
Prior to this change, when no user was logged in and the AI Setting was
queried via the GraphQL endpoint, the query resulted in an error. The
missing user resulted in a nil object in the AI Setting policies.

This change fixes the error by adding a safe operator to the AI Setting
policies.

Changelog: fixed
EE: true