Check for setting before showing ImportHistoryMenu
We should check that the application setting for bulk_import
is enabled or if the feature flag override is enabled, before
showing the Import history menu  to users.

This is similar to the controller that renders the page which has
the `ensure_bulk_import_enabled` check.

Also add specs to cover the new (more complex) logic.

Changelog: fixed