Add group_users, test all associations
In this commit we add group_users as it will be needed in the eventual
approvers method in this class. The behavior is inspired by the code in
the ApprovalRuleLike concern.

This commit also adds tests for the existing associations. The tests are
more detailed for group_users because it has special logic like
`distinct` as part of the definition.

Some of the testing code requires a sharding key to ensure the test
records are valid.