Put examples in block quotes
It's hard to differentiate between the instructions
and the examples, so we should put rendered examples in
block quotes to give a visual identifier of rendered examples.
We can also introduce every rendered example, and skip links
to the raw file for examples that work in the main docs site
already.