Remove deprecated Terraform CI/CD templates
This change set removes all Terraform CI/CD templates and its associated
documentation. This is a planned breaking change for the 3rd breaking
change window according to
https://docs.gitlab.com/update/breaking_windows/#window-3
and announced in GitLab 16.9 with
https://docs.gitlab.com/update/deprecations/#deprecate-terraform-cicd-templates.

Once the templates are removed, existing pipelines using one of the
removed templates will fail during pipeline creation. The alternatives
and migration paths are documented in the deprecation announcement,
which are to self-host the templates (the underlying images will still
be available) or migrate to the OpenTofu CI/CD component (recommended).

This breaking change is a security exception because we can no longer
ship updated Terraform binaries with our templates due to HashiCorp's
license change.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/438011

Refs https://gitlab.com/gitlab-org/gitlab/-/issues/438010

Changelog: removed