Fix sorting of MR diffs stream
There's an issue when listing MR diff files when fetched from
Gitaly using `MergeRequestDiff#diffs` with `offset_index` as it
was returning files in an incorrect order.

In order for us to fix that, we get the files from diff files from
`merge_request_diff_files` table as it's already sorted. We get
a list that's already considering the `offset_index` provided.

We use the paths of those files when fetching diffs from Gitaly
and then we re-order them in Ruby when we call `#diff_files` with
`sorted: true`.

Rapid diffs is still behind a feature flag.