Skip resolved_by validation on importing records
A user on a GitLab instance can resolve a note and subsequently be
deleted. This results in a note with a `resolved_at` date, but no
corresponding user.

This note is then imported into other GitLab instances with
`resolved_at` intact but no `resolved_by_id`. This fails validation, and
the notes fail to import.

Since the invalid note is allowed on the source system, i.e. nothing
prevents us from deleting users and leaving "orphaned" resolved notes,
I think it makes sense to skip the validation on import.

Changelog: fixed