Add default value for canManageSelfHostedModels injected prop
The GitlabDuoHome component is shared between the SM instance settings
page and Saas group settings. However canManageSelfHostedModels is a
needed check only for SM instances (self-hosted models do not exist
under group settings). This means the prop should have a default value
on it otherwise it will create an error when it is not passed in during
sass mode.