Document SaaS-only GraphQL schema
Previously SaaS-only GraphQL `ServiceType` enum values were absent from
our documentation, although they would be present in the actual GraphQL
API on GitLab.com.

This change allows SaaS-only schema to be present for when we compile
the documentation, but not present in the actual GraphQL schema.

The Saas-only `ServiceType` enum values are described as
"(GitLab.com only)".

Changelog: changed