The base implementation of the PDF Vulnerability Report
Context
---------------

We have an [epic][0] to build out PDF reporting for our
`vulnerability_management` features.

We previously built a proof of concept, which can be referred to for
the decision to use `prawn` vs other pdf generation strategies. [^1]

This Change
--------------------

The goal of this change is to get all the plumbing in place, so we can
start getting pdfs generated and in front of design, to help iterating
on the UX and UI.

To that end, this change:

- Adds the skeleton for the specs for individual pdf UI
  components (currently just the header)
- generates a pdf (containing only the header)
- provides a patch (on the [MR][3]) to add a placeholder export
  button in the UI.

[0]:https://gitlab.com/groups/gitlab-org/-/epics/16989
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/183327
[2]:https://gitlab.com/gitlab-org/gitlab/-/issues/517980#poc-outcomes
[3]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185058

[^1]: The proof of concept [MR][1] builds out a demo using the
      gems. The [issue][2] has a writeup on the outcomes of the proof of
      concept

---

epic: https://gitlab.com/groups/gitlab-org/-/epics/16989
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185058
resolves: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185058
Changelog: added
EE: true