Hide disable PAT group setting on self-managed
The ability to disable personal access tokens for users of a group
is limited to GitLab.com with enterprise users. This setting is
not functional on self-managed but was shown to users, causing
confusion. This change ensures the setting is only shown for
GitLab.com licensed groups.

Changelog: fixed
EE: true