Add table for Organizations::OrganizationUserAlias
As part of allowing per-organization internal bot users, we need a way
to reference these users by the aliases that have been used for the
legacy cell so the bots work consistently across organizations. Adding
this table for per-organization user references is the first step in
this process.

Changelog: added