Log errors received from Anthropic API
We can receive error responses from Anthropic API especially
when we go over the context window limit with our input. In order
to confirm that they are indeed due to context limit, we should
log the errors.

This also includes logging of total tokens used for successful
requests (input + output).