Return correct export class for pdf vulnerability exports
Context
-------------

In an [earlier MR][0] I began adding the foundation for a PDF
vulnerability management report.

While working on the next issue in the [epic][1], I have noticed a few
additional followup tasks that are needed

This Change
-------------

This method is only used in the `SegmentedExport` workflow, which the
pdf format doesn't currently use. However, it is a `public` method,
so I have updated it to correctly return the pdf service

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185058
[1]:https://gitlab.com/groups/gitlab-org/-/epics/16989

---

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/524055