Fix code dropdown for empty repo and activity project overview
In https://gitlab.com/gitlab-org/gitlab/-/issues/450667, we consolidated
the items under edit and code dropdown for repository/director view.
However, a few buttons are missing in the code dropdown when repository
is empty. This MR adds those back.

Additionally, adjust the logic for showing GitPod button to take
consideration of both user and instance preferences.

Finally, we no longer want to show the code dropdown in the activity
project overview.