Show backtrace when raising initializer DB connection
Due to the way we raise only after the yield has completed, we do not
know the DB was called from the routes file. So we use AR subscriber to
get the queries, and also the backtrace.

This will only log to the Rails log in test and development envs.