Allow agent CI access sharing across top level groups
Previously a project/group that was authorized to use the CI tunnel
had to share a root ancestor with the agent configuration project.

On GitLab.com this is typically not a problem, as most projects/groups
that would be authorized already belong to the same top level group.
However, self-managed installations often structure their hierarchies
differently with multiple top level groups, which limits the usefulness
of the agent.

With this change, any project or group on the instance can be specified
in the agent configuration file, regardless of root ancestor.

This functionality will not be available on GitLab.com, and is
controlled by the `organization_cluster_agent_authorization_enabled`
application setting.

Changelog: added