Solidify workloads abstraction
This work is expanding on the new `Ci::Workloads::Workload` model that
is being introduced to support Duo Workflow and a couple of other use
cases. The workload model was originally introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/176742 and the
idea is that this model will allow us to run arbitrary workloads inside
of a CI Pipeline on demand. For the time being it only abstracts the
caller away from the `::Ci::CreatePipelineService` and the details of
constructing a valid `.gitlab-ci.yml` string.

The reason this abstraction exists is that we see demand for various
ad-hoc workloads in many features in GitLab (see
https://gitlab.com/gitlab-org/gitlab/-/issues/328489 and
https://gitlab.com/gitlab-com/content-sites/handbook/-/merge_requests/10811
for more examples) and today it's easiest for us to build these features
using the `Ci::Pipeline` as the execution environment. There are longer
term architecture opportunities (e.g.
https://gitlab.com/gitlab-org/step-runner/-/issues/78) for us to have
different ways to execute these arbitrary workloads. But having a lot of
tight coupling in the codebase to pipelines and CI job yaml is going to
make it hard to swap out this internal execution environment longer
term. Aside from simply providing the ability to easily switch execution
environments as our architecture changes it is also probably more
maintainable to abstract developers away from all the complexity that
comes with CI pipelines. The abstraction also helps with security as,
for example, we are able to strip all unnecessary CI variables from
workload jobs.

For reference we can look at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/182840 for how
this workload class may be used in more parts of the codebase in future.
But there is no immediate need to migrate all those things to workloads
just yet so instead we are focusing on the minimum needed to unblock Duo
Workflow.

This MR makes some tweaks to the abstraction:

1. It introduces a `WorkloadDefinition` which is just a wrapper to
   prevent consumers from having to pass around a job Hash everywhere.
   Internally it just stores a Hash which describes a CI job
2. It introduces a `SimplifiedWorkloadDefinitionBuilderService` which is
   a builder class that allows consumers to set the various parts of the
   workload they need (e.g. image, commands, variables) without needing
   to know how to construct the CI YML hash. This class constructs the
   correct CI job definition from all the various settings provided
3. It updates the `RunWorkloadService` to pass in the newly created
   branch name to all workloads as `CI_WORKLOAD_REF`. Note that this is
   necessary in workloads because we skip passing most of the normal CI
   variables
   and that means we do not have `CI_COMMIT_REF_NAME` available in our
   jobs. Additionally this protects us from becoming coupled to the
   specifics of GitLab CI pipelines/jobs by depending less on the full
   set of predefined variables
4. Updates our Duo Workflow related code to use this new API