Add exclusion list for false positives
I finally found a false positive for a helper method that was identified
as "unused" by this hacky little script, so it is time to start building
out an exclusions list to avoid these when scanning. This is a step
towards hardening this script enough to add it to our regular scanning.