Introduce the test_already_failed_on_default_branch failure category
This is to track backend tests that already failed on the default branch.

We detect those by looking at whether we got an 112 exit code in the
script portion of the CI/CD job. When we do, we create a file that we use
as a marker for the after_script (we do this because the exit code is not
accessible from the GitLab job API resource).