Refactor AI Setting mutation to return the object
Prior to this change, the update mutation for the `Ai::Setting` object
was returning specific fields. This made it more complex to add
authorization logic for the fields.

This change refactors the mutation to return the `Ai::Setting` object
instead of specific fields of the object to simplify the authorization
logic.

Changelog: changed
EE: true