Allow auto enable Duo Core feature setting via seat link sync
Prior to this change, the Duo Core features setting were always auto
enabled whenever an instance was activated. This didn't cover auto
enabling the setting for early activations of future dated
subscriptions.

This change adds the auto enabling of the Duo Core features setting to
the seat link sync to cover the early activations of future dated
subscriptions. At the same time, the shared logic for updating future
subscriptions, provisioning add-on purchases and updating those setting
are moved to a dedicated service to maintain the logic in one place
instead of in two places.

There's also a change in behavior when updating the setting. It is only
updated when the current value is `nil`. It will no longer always be
auto enabled during activations. This also align the auto enable
behavior with the one for GL.com that didn't have a similar entry point
for a subscription start like the SM activation.

Changelog: changed
EE: true